LEAN-3647 ICR-030 - Move the Section Category selection to the body edit…
