+++ src/configs/editor-plugins.ts
@@ -38,8 +38,8 @@ import objects from '../plugins/objects'
 import paragraphs from '../plugins/paragraphs'
 import persist from '../plugins/persist'
 import placeholder from '../plugins/placeholder'
-import section_category from '../plugins/section_category'
 import section_title from '../plugins/section_title'
+import section_category from '../plugins/section-category'
 import sections from '../plugins/sections'
 import selected_suggestion from '../plugins/selected-suggestion'
 import table_footnote from '../plugins/table-footnote'
